fix: handle last word gracefully
Normally I don't allow words to be revalidated once the guess was wrong.
But this approach doesn't work with last words, which need revalidation
otherwise the flow can't be ever finished.

This fix also makes it possible to guess just one translation, which
isn't very useful but ofc improves the correctnes of the program.