add: prevent accent from making a guess invalid
Accents are important for written language to a degree but not terribly
important overall. That's why I only want to notify user that the
accents don't match but leave the guess valid.

The notification part isnt' done here yet, but I think a warning in the
tester area should be enough.