change: remove partly matching words
This was used so that eg. plural forms don't make a guess invalid when
in singular, but this case is rare or can be prevented by a correct
original <-> translation definition.

I already met with cases where I guessed a word only partly, which in
effect made my guess logically invalid by this partial matching marked
it as valid anyway, and this commit prevents this situation.