add: finish feature
It happens every so often that I have many words to go through but not
much time. So I create a simple finish button that just takes all words
I didn't go through, marks them invalid and so allows them to be saved
for later. The missing bit is that they are not saved yet so one has to
copy them immediately.