refactor: put try again into Results
Before, the try again and reset was present in the Tester, which
logically nor architecturally made sense. So moving it to Results was a
logical move.

Also, I introduced a done state to the app since inferring the done
state from other state like words wasn't scalable, and it turned out a
dedicated done state is a must.