feat: make config persistent
Finally I got so pissed with config returning to the previous state that
I decided to make it persistent. The mechanism of adding a new config
prop is not very easy, since it has to be hooked up with the app state,
the persistent storage and has to initialize a toggle. But it's ok and
moving this responsibility to a dedicated Config component helped a lot.