change: peeking logic
Previously peek wasn't shown when translation was invalid even though
peek=true. I removed the `valid` condition but it has the negative side
effect in the vocabulary view that when a word is guessed correctly, it
doesn't automatically show the peek. That's fine since WriteTester is
too low level to decide this imo but it needs to be fixed in the
Vocabulary view.

Also, I did a little refactoring to make the WriteTester reusable since
it's a very handy component.