add: language
The language is already processed on backend as well.

It's distributed among child components using context. The interesting
thing about context in solidjs is that values passed into it are not
magically made reactive. That means if I pass a simple string into a
context and change it later to a different value, child components won't
get updated. That means that values entering context have to be reactive
already.