fix: put mood back into tense name
The thing is that multiple moods can have the same tense, and this led
to conjugationTester having double the fields. It happened in Spanish
when I selected Indicativo mood with Present tense. However, the tester
would display also second set of fields because Subjunctive mood also
has Present tense, and because I actively removed mood from tense name
in the past, my app logic couldn't tell that one belongs to Indicativo
and the other to Subjunctive. So I returned the mood to tense's name and
only display the moodless version in the Filter.