fix: conjugation tester rerendering
What a ride! Let's start from the beginning. I had a problem with
selecting additional tenses after I already wrote something into
ConjugationsTester - the conjugations got reset and my text disappeared.

At first I blamed WriteTester's onMount which indeed sets input's value
to ''. But onMount runs only once, so this couldn't be the problem. But
before I realized this, I embarked on conditioning it only to when props
have changed. I was looking for something like `onProps` which would be
able to compare new and old props, but apparently there is no such
thing. Before abandonning this idea, I resolved simply on storing the
old props manually in the component's scope.

After that I realized that WriteTesters actually shouldn't rerender at
all when I add a tense, since the current tense stayed the same. I
tracked it down to <Show> causing the needless rerender, particularly
because of the `keyed` attribute. Now this attribute has been a source
of confusion to me, also because the docs don't explain it at all. Also,
I confused myself somehow by thinking that only when keyed=true the
`<Show>` can accept a function as children, which isn't the case. But what
does `keyed` do exactly? It doesn't have any effect on the `<Show>`s
primary function of showing children, but it says more explicitly whether
the children should be kept or replaced, so essentially like React's
key. By default, with keyed=false, the content inside would be
considered the same no matter the value passed in to the `when`
attribute, so it wouldn't rerender, which is something I actually desire
so I chose this path. But with keyed=true one has to option to specify a
key which when it changes, the `<Show>` children will get replaced. And
this is determined based on a strict comparison.