refactor: simplify the settings component
Previously, adding a new setting property required to change code in
about 8 different places just to propagate it to the consuments. Which
clearly exposed the bad design.

First of all, I was storing it by property in both the memory and
storage. This needlessly complicated adding new properties and I decided
to store the whole settings collection in an object. This lead to
simplification of both the component's and settings' API.

Lastly, I also decided to set the `checked` property in Toggle from the
outside. There can be valid cases where the Toggle needs to be rendered
but it's current value will be changed later on when data loads.