add: option to remove word from vocabulary
I found thanks to Andrejka that all the plans of clever algorithms,
telling user which words are ripe to be removed from vocabulary, are
actually too complicated to implement and probably not needed, by doing
a simple thing - just letting users to pick themselves which words they
are sure about and don't want to practice anymore. So I did just that.

Also, I somewhat improved the experience in the VocabularyTester by
autofocusing the input field when Next or Remove is clicked.