add: validate conjugation on next/finish
It turns out people actually expect the table to be validated when next
or finish is clicked, which is understandable.

So far I am kinda harsh with it since I don't allow the user to proceed
until all are valid. I should create some easier way out, but for the
simplicity's sake in this commit I'll keep it this way.

Also, I started using a not so obvious pattern, and that's propagating a
kind of child component instance to parent. In SolidJS, there is no
obvious way to call methods on a child, and they have to be exposed
through props manually.

I also found an interesting pattern where this get's exposed even earlier
in child's lifecycle like this
`const { Modal, openModal } = createModal();`
which is a great way when such a component is more or less static in
parent's template, but I don't think is very practical with more dynamic
ways like when there are multiple in a list.