change: improve vocabulary view's layout
Basically, I made the VocabularyView use the whole space given to it,
and as a result I can afford to put the VocabularySettings to the bottom
edge in order to declutter the view.

Please note that the app is still unusable on smaller phones and I'll
have to make the WriteTester and the related components responsive,
otherwise content will overflow.