add: basic conjugation results
I show tables for each mood+tense each with expected and actual results.
Soon I will add an icons to it to make it apparent which were correct
and which incorrect.

I had to remove the condition completely that user has to have all
conjugations valid before moving on, since it doesn't make sense, but I
should add some conditioning on when exactly she can move on. Eg., if
she pressed Next and the form wasn't valid, I should only validate the
first time, but allow her to proceed on second action so that she
doesn't get stuck.