change: reset conjugation view when different verb applied
Previously applying a verb would have no effect on the mood/tense filter
and user would A) get no feedback on whether the new verb has already
loaded and B) have a hard time to continue using the conjugations view.
So I reset the filters and the view's state in general.

One interesting thing is that whenever I find out that a certain child
component is not following the suite in terms of it keeping its old
state, it's always caused by the child keeping that state unnecessarily
in the first place. In what feels like 99% of cases I am able to solve
this problem by hoisting the state into the parent. As a result, when
parent's state changes, it's enough to change all the children.

This might seem like polluting parent with too much state on the first
look, but implementation-wise it's so much easier - I just move the signal
into parent, add new prop to child's props and expose callback to set
the state from the child. Alternatively, if I were to reset the child's
state from above, I'd have to propagate child's reference to the parent,
create callback for that purpose since it's not automatic like in
Angular, and call the reference's method while also still keeping the
state in the child. And of course, that would mean I'd have to make sure
the two are always in sync state-wise. That's why I am pretty sure state
hoisting is the solution most of the cases. The only exception is
reusable components that have a good reason to maintain their own state.
This might be due to them to a great degree being independent of their
parents and context.