add: practice only incorrect conjugations
Previously the whole tense had to be practiced but this is tedious so I
only make the user practice the incorrect conjugations. This leads to a
pretty cool experience imo.

Also, the state in ConjugationsView has grown quite a lot, and I think I
should at least organize the various signals into one createStore.