change: hide tense filter when practicing incorrect
I did it so so that I don't have to deal with the exotic case where user
decides to pick something in the tense filter while also practicing
incorrect conjugations. This would create a complex situation which I
don't want to deal with right now, so I just forbid it altogether by not
showing the tense filter.