fix: use conjugation validators correctly
First of all I call them manually when user submits a conjugation table,
and secondly I clean them directly when next conjugation table is being
prepared, or when the current one is the last one.

Before, I was using createEffect to clean up the previous validators,
while the ref callbacks in children - WriteTesters - were refilling the
validators with the next ones. But to my suprise, the effect was called
only after the ref callbacks, so I chose this more direct way, which I
think is actually much better.