add: make tense filter list expandable
In order to decrease clutter in the UI. I was previously thinking about
clever, CSS only solutions, but white-space: nowrap with ellipsis was
still displaying the checkbox's box and other solutions didn't seem to
offer good options.

I also considered hiding everything beyond the first line and reading
the list's scrollHeight, which would detect there is a hidden content,
and based on it offer the option to expand, but that's too much JS for a
simple task, and I'd have to handle resizes and shit.

So I settled for a simple slice(0, 4) on the tenses.