fix: focus on first conjugation tester field
Interesting problem! I relied on the following pattern: WriteTester uses
ref on its input to call a function, which receives the input DOM node
and sends it to the parent, which in turn call focus() on it. Sounds
logical, but it didn't work!

I then found out that the DOM node is not yet attached to the DOM at
that moment. Which means the ref is triggered as soon as the node is
created and doesn't wait for the attachment.

That's why setting the ref, but giving it to the parent to call focus()
on it has to be done only `onMount`, when the whole WriteTester
component gets mounted.