fix: clean signals in the right order
Something I have felt for some time as being bad design actually
backfired on me! ConjugationsView has many signals and some are
interconnected. Take `conjugations()` and `selectedTenses()` which are
both used to derive state for children components. I have known for some
time that having these as separate signals didn't feel right. And now
this fear became real - it turns out that the order, by which these are
set actually matters! And it was that order that broke ConjugationsView
when:
1. user chose verb to conjugate
2. selected mood and tense(s)
3. searched for another verb

Which lead to TypeError because the `conjugations()` were emptied while
`selectedTenses()` still kept asking for particular tenses in the
former.

This is bad, very bad, as Trump would say. The quick solution for now is
to change the order, but the proper solution will be to use
`createStore` and set changes in one go, not separately.

https://trello.com/c/13HczJff