fix: shallow copy initial store state
I completely missed that initial state is just an object, which if
consumed by store will be made reactive without copying. Which somewhat
makes sense because the store has to make the object's properties
reactive as well, so it'd have to create a deep copy, but that's not
mentioned in the docs at all, so I shouldn't have expected it. As a
result, consuming the initial state directly lead to it being changed
in place, therefore polluting it, and breaking any kind of logic that
relied on fresh new state for reset.

That's why I decided to use `getInitialState` function instead of an
object. While the former forces new object copy each time it's called,
with the latter I'd always have to remember to shallow-copy it by hand,
or face unexpected bugs.