add: make chips focusable and tabbable
I thought this was gonna be a boring little task, but was I wrong!
Instead, it taught me an important lesson about reactivity.

The trigger was me turning the chips into buttons in a list. The became
tabbable and focusable, so far so good. But I noticed that toggling a
chip somehow made it loose focus, which got reset and moved to the whole
list.

Soon I identified that the whole list got rerendered and since the DOM
nodes were replaced as a result, logically the focus got lost. But why
was it rerendered in such a forceful way? The answer was evident, the
`<For>` was using a computed signal which, if anything changed inside,
produced a new array of fresh new chip items. But if such a common case
lead to unnecessary list rerenders, how is this solved correctly in
Solid?

Naturally, I began looking for a solution equivalent with React's key,
since that's its solution to this very problem. And I quickly found
there is no such thing in Solid. Instead, I found this article expanding
on the same topic: https://github.com/solidjs/solid/discussions/366. What
it says is that instead of recreating the whole array of items, one
should trigger the one nested signal that changed instead! This is
really powerful, since it doesn't require any diffing to find real changes
in the markup, but rather directly targets the changed place through the
concrete signal.

This means reactivity doesn't require us to do a bunch of object copies,
but rather allows us to work with the existing ones in a very targetted,
efficient way. I like this way of thinking very much, but it some
getting used to indeed.