test: add first test!
Originally I wanted to build in Jest, but that turned out quite a pain.
There was a ton of presets for Babel, which while bad on its own wasn't
still as bad as having another transpiler in the codebase just for the
tests.

Despite that I committed to it but failed to even run a simple
test. First of all I was presented with hurdles caused by outdated
solid's template for testing, which I should update btw. Secondly, I
just didn't get Jest to stomach ES6 module syntax, even though I
installed and configured the typescript preset.

Frustrated, after more than an hour, I dediced to look for an
alternative and found vitest. Its philosophy is to take the existing
production configuration - my vite config - and use it as a basis for
tests. How refreshing, logical and the total opposite of what Jest is
doing! And indeed, this simple logic allowed me to run my first test in
under 5 minutes!