add!: introduce solid-ui
The main reason I introduced it is that I don't want to have to deal
with the intricacies of component development. It was cool to try out
what HTML has to offer, and in the beginning where I had just a few
components it kinda made sense to just use HTML only, sprinkle a little
CSS and JS on top and by doing so limit the dependencies.

But as the scope of this project is growing, as well as the number of UI
widgets I have to create and support, creating these UI components and
putting them into a unified design framework is a black hole for my time
and frustrating experience at places too.

So the solution is some kind of a design system with out of the box
components. This would speak in favor of a component library, but since
I want to stay open to greater customizations, I decided for solid-ui, a
port of shadcn/ui to Solid. It allows me to set up a design system,
generate components and customize them heavily if needed. So it removes
quite some burden off of my shoulders.

It's not perfect, and I might end up using Tailwind UI or Flowbite, but
for now I'll got with solid-ui.

This is a breaking change since all the previous styles got purged,
which is ok because the styling wasn't great before, but I'll need to
rewrite all views to use the new components.

I also went in direction of both light and dark mode, since why the hell
not, I like both.

In case I decide to switch away from solid-ui, reverting this commit and
getting rid of the generated components should suffice.