add: solid-ui components
They are beginning to replace my own, eg. Input and Dialog got replaced
by this superior, but arguably also a little chaotic implementation. I
say chaotic since it's not always easy to find out which props can be
passed to the components, but that's the price for their flexibility.

I still have to convert the actual vocabulary tester and the whole
conjugations view, but I'll get there naturally.