change: limit vocabulary CSS size
So that it fits the screen better. I am not happy with defining the
number of columns for each screen breakpoint, and I should probably set
each vocabulary item to a more or less fixed width, and let flex-box
take care of the rest. But it's not a priority now.