add: initialize vocabulary testing by navigation
It turns out solid resource is amazing! It's a natural caching mechanism
and as a single object with the possibility to either mutate or refetch
is so much simpler than Redux and similar solutions.

I also continue to keep router stuff outside all children components.
Let's see how long this lasts.