add: words editor
I had to abandon the onBlur edit because it just doesn't work with the
Sheet primitive, which is actually a Dialog. The problem is that when I
edit a field in the sheet and then click outside to close it,
surprisingly the blur event doesn't happen on the field. It seems like
it's not permitted to lose focus when clicking the backdrop. No matter
what I tried. I tried blocking the pointer down custom event which
prevents the Sheet from closing, and replacing it with a click instead.
That makes to me more sense anyway since pointer down is premature
marker of user action in my opinion (click can be cancelled).

So instead I batch the updates and make user confirm them before closing
the Sheet.