chore: get rid of old checkbox
Unifying the UI by using the solid-ui's checkbox made the old Checkbox
superfluous.

One interesting thing is that the generated checkbox didn't include any
label. That responsibility lied with the user, but I find this
impractical, so I put it in. Let's see how it fares. If only in some
cases I'd need custom label, I can still hide the build-in one and use a
custom one.