change: improve the lang picker
Uses solid-ui's select for it. I am not too happy with the select
though, as it's pretty complicated to put together. One needs to use a
special trigger, value, content and item elements and all wrap it with
top level select. Not exactly easy usage. It's quite possible that I'll
need to abstract these away.