add: make a combobox out of country select
There is one hurdle with the triggerMode. Currently, the combobox only
opens when user starts typing or click the chevrons icon. I wanted this
to happen once user clicks on the input field, but Kobalte only offers
focus. The problem with it is that it doesnt' close the dropdown list
after user selects, as described here:
https://github.com/kobaltedev/kobalte/issues/387