add: show current route in navbar
This is done using 2 interesting approaches. First of all, I started
using `<A>` from solid router. This is the same as native anchor, but it
provides active and inactive CSS classes. Those I wanted to use to show
active route.

Second problem was how to utilize them, since I only use Tailwind. For
that, I use css module and I make nested style rule global by using
`:global`, so that I can target these classes.

Last problem was that when going to /, it would show VocabularyOverview
component, but the navigation Vocabulary `<A>` href wouldn't match the
current route, and active class woudln't be assigned. For that I use a
`<Navigate>` component in the routes config, which immediately navigates
to given url, and that allows me to match the route with the `<A>`.