fix: update words in editor when new added
This problem was actually bigger than I thought. First of all, I wasn't
mutating the vocabularies resource properly as I completely left out
newly created items. This made sense since I was calling
updateVocabularyItems, where items already had IDs and therefore didn't
need to be refetched. This is not the case with newly created items, so
I needed to create createVocabularyItems to do fetch and proper mutation
on the resource.

After that, I found that even with this solution, the list in
VocabularyEditor is still not updating. It came down to the Search since
it's responsible for what gets displayed in the former. But it wasn't
updating its search results when the search terms changed. Redoing
search upon receiving new terms solved the problem.