test: search in VocabularyPage
I did this just by taking the component in question and using the
already set up infrastrucutre around mocking apis. The only other thing
I needed to do was adding a MemoryRouter which is ideal for testing
purposes.

I got to say that I quite like this setup, but at the same time I don't
want to overdo it with these kinds of component/integration tests
because I imagine those will be pretty expensive to maintain.