fix: make vocabulary tester always submittable
This is an interesting one. The vocabulary tester was submittable only
when it was invalid, not when valid. It turned out this was due to the
usual submit button being disabled, while the "next word" button was of
type "button", so there was no submit button present and pressing enter
didn't work. Now, I make sure there always is a submit button present.