add: save progress on each word evaluation
I was thinking whether to save progress on unload or beforeunload, but
this is a suboptimal solution. First of all unload is deprecated and the
latter leads to deoptimization with the bf cache. Secondly, such events
firing is not guaranteed, so the user might lose progress.

In the end I decided to save progress after every word evaluation. This
is not an optimized solution DB wise once I start storing it there, but
even then I could queue the updates and send them in batches.