test: the vocabulary creator submit
Using HTML apis is nice but there is no type safety. In the last commit,
I was fixing this exact issue. I had a wrong name in the form element and
so this lead to the form submission not passing the validation.

I couldn't use the `screen` from testing-library because the combobox is
hard to identify, and testing-library has only limited ways to query
elements. This is by design btw because they want you to query elements
as if the user has seen them. I had to use querySelector in the end but
clearly, this should be improved.