test: improve VocabularyCreator testing
I was relying on querySelector before since I wasn't able to query the
country-select properly. I however found a way to make it work. The
problem with my previous country-select was that I was putting the id on
the combobox container, but its label can't reference it if it's not a
form element. Which it wasn't since it was a div.

So I moved the id, and the name for that matter, to the hidden select,
which fixes the labeling but also still leaves the native form submission
mechanism intact. This allowed me to use the testing-library's screen
to query all my elements.

Also, this allowed me to write the test in a more declarative way by
simply selecting by a visible text like a label. I am not sure this is
always a good idea, since it makes tests less durable since the text can
change even though the logic doesn't have to.