change: upgrade solidui components
One lesson I learned is to keep the modification to components to a
minimum, because otherwise they will be hard to upgrade once kobalte
releases updates. On the other hand there are bugs/improvements I should
get into the solidui, like:
- select and combobox scroll overflow for long dropdowns
- dialog scrolling on small height viewports

It can be there are small bugs in the UI but overall I am glad I did
this upgrade, since it seems to have brought many small changes. And it
brought bugfixes as well, like for the combobox trigger mechanism w/ the
value "focus", which didn't work properly before. It also made me aware