fix: keep tester validation state on word update
When word was updated previously after performing validation, it would
perform it again and make it invalid. That was because of the effects
in the WriteTester. The problem was that I wasn't distinguishing between
new words and the same updated words. Now I am checking word IDs to
prevent that.

Also, I found in the VocabularyTester that I was filtering words left
based on their original string, but that's of course wrong when I allow
the words to be updated. Again, using IDs solved this.

I also made the realization that createEffect's callback's return value
doesn't determine when it's gonna be called, no, it only supplies the
previous value for the cb. The cb rather runs when any of its reactive
deps change. Good learning there :+1:.