add: simple dashboard page
There are few caveats in the implementation. First of all there is only
a dummy video. Secondly, the fetching of the recent vocabularies isn't
connected to the primary vocabularies resource, although the argument
can be made it doesn't need to since it fetches only a limited range.
Thirdly, it doesn't even try to handle anonymous users, but that's
something I will take care of in the next commit. I wanna simplify and
hide the whole app behind a login screen.