add: hide app behind auth guard
Adding new pages which require user to be logged in showed how bad the
current architecture of letting anonymous users in the app is. It doesn't
scale and it complicates the component code with conditional rendering.
That's why I ditched this paradigm and am now hiding the whole app behind
the auth guard.