change: auto peek after wrong guess
This makes the testing flow much easier:
1. User guesses
2. Guesses right, can press Enter to go on. This already worked.
3. Guesses wrong, and this is new, can press Enter too to go next.
4. If guessed wrong, gets the auto peek to see what the correct answer
was.

In the future, this'd be nice as configurable.

I also fixed the submitability of the 2 buttons in the WriteTester. They
should be much more intuitive to use now. Before this, the skip button
would not be submitable and Enter couldn't be used to go on.