fix: prevent broken tester when switching vocabs
The problem was that when switching vocabs, the underlying vocabulary
resource would take some time to load the current vocabulary. And the
components just weren't ready for this. First of all I kept 1 instance
of VocabularyTester around for multiple vocabularies, which is just too
risky. Secondly, I wasn't handling the case when vocabulary is loading
at all which lead to wrong vocabulary being passed to the tester at
first.

That's why I added a check for resource loading and am forcing a new
instance of VocabularyTester to be created when switching vocabs via the
mysterious `keyed` prop, which thanks to this I understand again a little
better.