refactor: create a transform layer between api and resource
This commit is a defensive measure. Previously I used minimal transform
between the objects returned by DB and those used in the app. This of
course not only creates a direct dependency between app logic and DB
objects shape, but it also makes it awkward to make the two differ,
wherever this makes sense.

So I created a layer of transform functions, kinda like the transformation
we are doing in services in the Slido Upgrade. It feels a bit heavy
handed, but I like these things about it:
- The api is no more responsible for transforming the objects. If I kept
  it the old way, I would sooner or later have to introduce some kind of
  application logic there and I don't want to do that.
- It's also clearer now who is responsible for transforms - the resource
  layer. This allows me to add some logic or merge more objects together
  in there without feeling guilty.

It might turn out this is an overkill, but I wanted to prevent further
code rot.