add: video to the auth page
I added this video to auth page so that people can see what the app is
about before signing up.

I also removed `<main>` from App.tsx because that's not how main should
be used. It's not a container that I want to lock down to whatever page
gets displayed inside it, it's a specific marker that can differ page by
page.

Also, love the fact I could use `<aside>` for the video!