fix: vocabulary words sorting
I was sorting the localized date-added strings, not the time milis, and
that created a wrong sorting order. I use the milis now. Also, I use Map
instead of an object since object permits only strings as keys, but I
need numbers. Not impossible with object, but I don't want to bend it.