fix: prevent resetting test on word edit
The problem was the "ingenious" use of the keyed attribute above the
VocabularyTester. When the `words()` changed which happened on word edit,
the keyed kicked in and recreated the tester. This was of course not
desired in mid-test where the progress ought to be kept.

I chose not to test it right now since that would require testing the
page component with router and everything, and while I think it's
important I just don't feel like spending much time on this now.