fix: reset app on logout
The previous problem was that switching accounts would not reset the
vocabularies resource, so vocabularies from the previous account would
be visible. This was fixed by resetting the app on logout.

But this commit is about more than just that. I refactored the resources
and especially the vocabularies resource. I moved the initialization of
the resource to the `init` method instead of the get function, and kept
it lazy by accompanying it with a signal. This signal, since it's only
a trigger, is of type boolean. Initially set to `false`, which is one of
the values that doesn't trigger the resource fetch, is set to `true`
once fetch is in order.

All in all I am very happy. I solved the aforementioned issue and
refactored the resources in such a way that I can read and trust them
now and have enough confidence in them to expand the architecture.