change: use 2 col layout in vocab tester
Some users had a problem with the settings being at the bottom of the
page as they were easy to miss. So I decided to go for the 2 col layout
and put the settings into an aside.

This has the benefit of not having to use grid, and in fact I removed it
from the App component as well. The usage in the App component was wrong
since it assumed structure of its children (rendered routes) and there
was this indirect, implicit dependency between the 2 in terms of layout.