fix: vocabulary switching in VocabularyTestPage
I didn't account for stale vocabulary in the VocabularyTestPage. This
means, that when I left the tester and picked another vocabulary for test,
it still took the old vocabulary for testing. But then the new arrived
and it messed the tester up. So I am guarding against the vocabulary
loading and delay the tester init until loaded.