add: imperatively submit add-words form on Save
Previously, user had to press the button to add a word before she could
save the added forms via another button. This wasn't a great experience
so I decided to make the "Add word" button press optional. For this, I
am using requestSubmit on the actual form object which, unlike `submit`,
performs validation and emits the SubmitEvent.

In order to do this, I had to do an overhaul of the WordsInput component.
The problem was, that it has 2 modes - raw and form, and both were very
different. The form one was a classic form using its own submit button
whereas the raw was just a textarea that was adding words, now to me
surprisingly, on blur only.

So what I needed to do was to wrap both modes in a form and use a submit
button so that the mechanics is consistent not only in the code but in
the UI as well. I even extracted the submit button to keep it DRY.

The lest step was to propagate the form element refs upwards which
didn't work in the beginning until I found that `ref` is a special prop
in Solid, unlike in React where it's not about props but rather about
the passed objects. This means that if I didn't use `ref` but eg.
`formRef`, the HTML element wouldn't be captured.

So now, whenever user presses the Save button, the words form is
submitted to stay on the safe side. If there was a valid input in there,
it doesn't get lost anymore, yay! The nice bonus is that it also serves
as a validation mechanism, based on the HTML5 validation attributes.
Meaning when there were no words added and the word form is invalid,
it doesn't let user complete the Save action and displays native
validation errors.