test: vocabulary create
I had an issue in aebe32f where the vocabulary to be created was
produced by the form just fine but wasn't transformed properly to the
DB object, failing due to TypeError. I had the flow fixed but not to
the point of transformation. So this is covering the whole flow.