add: confirmation dialog for deletion ops
It wasn't as easy as I thought in the beginning. The first problem was
that there is actually no imperative way how to close a dialog from
within. The reason I need it is that I don't want to place the burden
of tracking the open state in the client components.

So I thought I could use the Close button from kobalte's dialog primitive,
but I ended up not using it since it's not exported from the generated
'dialog' component by default. Sure, I can do whatever I want with that
component but after my experience from migrating kobalte 12 -> 13,
which meant many big changes, I decided in favor of DialogTrigger and
wrapping whatever button I want to guard with confirmation dialog.

The challenge now was to use both the trigger, which is a button by
default, and the button that triggers the dialog. I ofc didn't want to
end up with 2 nested buttons, but the `asChild` prop of DialogTrigger
turned out to be removed in kobalte 13. It turns out, that now all the
kobalte primitives support the `as` prop which can either specify a
native element or custom component. One has to be careful though to pass
props to the `as` component and put whatever handlers like `onClick` on
the trigger wrapper and not the child.

Now I love my implementation. Whatever button I want to guard, I just
wrap it in the ConfirmationDialog, yay.