add: storing of test results
This changes a lot. Up until now, I wasn't storing any results whatsoever.
In this commit, I am preparing the ground for persisting the test results
in the DB. For now, I am content with storing them in the local storage
since I didn't want to commit to the DB until I find out a bit more about
what I actually want from this feature.

A little surprisingly, it allowed me to remove quite a lot of code from
the tester components. I no more have to keep track of invalid and
leftover words, instead I store all and more in the test results,
including # of incorrect answers, skipped and leftover words. This
allows me to use the same format to save test progress, which actually
makes sense when I think about it. One of my tasks is to show results
when user interrupts test - this is a great precursor to that.

Few details to this commit. I initialize the results in the state of the
tester upon first render. This is not great but it's the same principle
I have already used there for `words` and the component is guarded by
the `keyed` attribute in the parent, so I am being consistent here (and
lazy ;)). I also gave up on cleaning the test results because they
actually don't matter in the meantime between tests. And last but not
least, I moved the init of saved progress to effect instead of onMount
since onMount runs only once and isn't reactive, unlike the effect.