change: fetch vocabulary progress in resource
Instead of in component. I spotted this and immediately thought it weird
to make a component responsible for that. The truth is that this was
just a makeshift solution. But, this changed prompted me to refactor
resources and app init as well since I'd otherwise expose a lot of code
to the indexedDB. So this refactor creates a vocabularyProgressApi and
mocks it in tests.