fix: vocabulary words controls alignments
Previously the controls were wrapping weirdly on smaller screens and
the system of only hiding the buttons but keeping the allocated space
created a lot of wasted space. So I removed all the fancy spacing and
just put all controls into 1 div. Once kinda fancy thing I did is that
I am only doing visibility:hidden on large screens so that there is no
shift on word selection but don't do such a thing on smaller screens in
order to save space.