add: move results to a separate page
The nice benefit of this is that I can go back to the results any time I
want, not only visit them after finishing test.

Also, this commit allowed me to do quite a clean up in the code - lot of
unused props and bad naming.

Last but not least, this task forced me to abstract idb-keyval away to
a new test result api. It still stays in the settings component, but
that's ok for now.