add: pie chart for results
Ofc I am using 3rd party for this - chart.js.

I removed the legend since it was needlessly taking up a lot of space
and will add my own the future.

I also fixed couple of bugs which were causing invalid words section to
be shown even if all results were correct.

For the future, I'd like to add some confetti, motivational feedback and
find a way to visualize the number of attempts for each word.